[#69] feat averageCRUD

+++ src/main/java/com/sparta/startup_be/utils/ConvertAddress.java
@@ -31,7 +31,7 @@ public class ConvertAddress {
         JSONObject rjson = new JSONObject(result);
         JSONArray items = rjson.getJSONArray("documents");
         JSONObject itemJson = items.getJSONObject(0);
-        CoordinateDto coordinateDto = new CoordinateDto(Float.parseFloat(String.valueOf(itemJson.get("x"))),Float.parseFloat(String.valueOf(itemJson.get("y"))),estateid);
+        CoordinateDto coordinateDto = new CoordinateDto(Double.parseDouble(String.valueOf(itemJson.get("x"))),Double.parseDouble(String.valueOf(itemJson.get("y"))),estateid);
         return coordinateDto;
     }
 
@@ -39,7 +39,7 @@ public class ConvertAddress {
         JSONObject rjson = new JSONObject(result);
         JSONArray items = rjson.getJSONArray("documents");
         JSONObject itemJson = items.getJSONObject(0);
-        CoordinateResponseDto coordinateResponseDtoDtoDto = new CoordinateResponseDto(Float.parseFloat(String.valueOf(itemJson.get("y"))),Float.parseFloat(String.valueOf(itemJson.get("x"))));
+        CoordinateResponseDto coordinateResponseDtoDtoDto = new CoordinateResponseDto(Double.parseDouble(String.valueOf(itemJson.get("y"))),Double.parseDouble(String.valueOf(itemJson.get("x"))));
         return coordinateResponseDtoDtoDto;
     }
